+++ swann.videostore/src/main/java/esgi/kata/videostore/Customer.java
@@ -23,28 +23,14 @@ public class Customer {
 		double totalAmount = 0;
 		int frequentRenterPoints = 0;
 		Enumeration rentals = this.rentals.elements();
+		// TODO move display to dedicated place
 		String result = "Rental Record for " + getName() + "\n";
 
 		while (rentals.hasMoreElements()) {
-			double thisAmount = 0;
+
 			Rental each = (Rental) rentals.nextElement();
 
-			// determines the amount for each line
-			switch (each.getMovie().getPriceCode()) {
-				case Movie.REGULAR:
-					thisAmount += 2;
-					if (each.getDaysRented() > 2)
-						thisAmount += (each.getDaysRented() - 2) * 1.5;
-					break;
-				case Movie.NEW_RELEASE:
-					thisAmount += each.getDaysRented() * 3;
-					break;
-				case Movie.CHILDRENS:
-					thisAmount += 1.5;
-					if (each.getDaysRented() > 3)
-						thisAmount += (each.getDaysRented() - 3) * 1.5;
-					break;
-			}
+			double thisAmount = calculateAmount(each);
 
 			frequentRenterPoints++;
 
@@ -60,5 +46,25 @@ public class Customer {
 		result += "You earned " + String.valueOf(frequentRenterPoints) + " frequent renter points\n";
 
 		return result;
+	}
+
+	private double calculateAmount(Rental rental) {
+		double amount = 0;
+		switch (rental.getMovie().getPriceCode()) {
+			case Movie.REGULAR:
+				amount += 2;
+				if (rental.getDaysRented() > 2)
+					amount += (rental.getDaysRented() - 2) * 1.5;
+				break;
+			case Movie.NEW_RELEASE:
+				amount += rental.getDaysRented() * 3;
+				break;
+			case Movie.CHILDRENS:
+				amount += 1.5;
+				if (rental.getDaysRented() > 3)
+					amount += (rental.getDaysRented() - 3) * 1.5;
+				break;
+		}
+		return amount;
 	}
 }