+++ swann.videostore/src/main/java/esgi/kata/videostore/MoviePriceCode.java
@@ -1,0 +1,5 @@
+package esgi.kata.videostore;
+
+public enum MoviePriceCode {
+  REGULAR, CHILDRENS, NEW_RELEASE
+}